Important update
Important update - various fixes

+++ cleanup.py
@@ -28,18 +28,15 @@ mylist = glob.glob("Camera_*")
 
 # Run through each folder and rerun script 
 for myfolder in mylist:
-
-	#curtime = time.time() - os.stat(myfolder).st_mtime 
-	#if (curtime/3600) < 1:
-	#	pass
-	#else:	
 	to_out = '%s /home/pi/Desktop/on_file_new2.py %s' %(python_interpreter,myfolder[7:]) 
-	subprocess.Popen(to_out, shell=True, stdout=subprocess.PIPE)
-
+	#subprocess.Popen(to_out, shell=True, stdout=subprocess.PIPE)
+	os.system(to_out)
 		
 # Run through each folder one more time
 for myfolder in mylist:		
 		
 	# If the folder still exists, then delete it
-	os.rmdir(myfolder)
+	curtime = time.time() - os.stat(myfolder).st_mtime 
+	if (curtime/3600) > 24:	
+		os.rmdir(myfolder)
